CopyFile() handles existing files better
Utilities.CopyFile() has been changed to be more resilient when the
destination file already exists. As the project currently is, there are
two scenarios we could run into this:

-When installing fonts
-When installing the cfg file

In the first case, if the fonts exist already, there's no need to do
anything else. In the second, this could happen if a new version is
being installed, in which case we would want to overwrite the existing
file. Even if they're the same version, there's no harm in doing so.

There is the chance that someone would run an old installer when they
already have an existing cfg file from a later install. However,
that's just a dumb thing to do. It's not worth the trouble of adapting
the code for this edge case when it's not a good use case anyway.