Changed the way config files are installed
Previously, the program would create a new "autoexec.cfg" file.
However, a much better approach (implemented in this commit) is to
install the config as "neodefaults.cfg", and write "exec neodefaults"
to the user's autoexec to make sure the neodefaults config is executed
whenever TF2 is launched. The program can more comfortably delete an
existing neodefaults.cfg file instead of an existing autoexec file.

This also supports Mastercomfig. The exec statement will be added to
user/autoexec.cfg if a mastercomfig preset is detected.