Updated error handling of install and logging
The install methods in Utilities.cs() have been updated to guarantee
that any Exceptions are handled without the program crashing.

Part of these changes include preventing the logging methods from
throwing Exceptions so that they may be safely used in catch statements
without nested blocks. This has the disadvantage of no reporting if
a message fails to log, but that is an unlikely event.