Allow logging to fail silently upon issues
However unlikely, it is conceivable that the logger could run into
intermittent issues. Since such issues wouldn't be worth crashing the
entire program (as it's not expected that anyone will really need the
logs), the program now explicitly declares that if any exceptions are
thrown while trying to initialize or write to a file, that request will
simply be ignored. This could cause confusion, but it is a trade-off.