Add support for custom/overrides file
In a scenario where a user personalizes the neodefaults.cfg file and
re-runs the installer, those modifications would be overwritten. To
avoid this, the file is now made read-only and a separate custom file is
provided where customizations can be placed. This file is executed at
the end of neodefaults.cfg. To make it clear these files are bundled
together, they are now installed under their own directory,
tf/cfg/NeoDefaults.

These changes also modified Utilities.CopyFile(), since it didn't make
sense to prevent the method from throwing errors when it is always
called in a try-catch statement anyway.