Enhanced log file handler
Two major changes were added. First, the way messages are written to the
file have changed. Instead of repeatedly appending multiple lines to a
file, Logger.cs will now assemble the message into one String in memory
first, then write the entire String to the log file once.

The other change is that a WriteLn() method has been added, allowing for
an easy way to add another line to the requested message. Since the
String is now assembled in memory, this will still be one write
operation and will be inexpensive.

Methods that were making a separate Write() call for the newline have
been updated to use the new WriteLn() method.