Enabled minimal HUD by default
The main reason for this is because bars like sandvich and cloak meter
are so ridiculously small in the regular mode that they're almost not
worth having. Minimal HUD creates a slightly larger bar for disposable
items such as these.

It also cleans up the display, leaving less obstructions. While this
does make health and ammo smaller and even harder to read, it is a
trade-off that is worth it.