Condense optional installations into one screen
There was no reason to have two separate bulleted lists on different
pages for additional customizations, so they've been placed together
onto one page, with check boxes to allow opting in or out of the
installs.

A help icon has been added to both of these to help give some detail
for each component, as well as a description pointing to the FAQ
for more information.