Add warning for disabling floating health
Using floating health is really bad because it requires the player to
look away from their crosshair to see their teammate's health. This is
particularly bad for medics. Thus, it is disabled by the config, but it
will prevent players from seeing the health of large bots in MvM. Update
the description with a comment so that any MvM players will be notified,
and can override this in their custom.cfg if desired.