Remove idHUD, and replace with custom HUD tweaks
idHUD was an important addition to NeoDefaults. However, that project
hasn't been actively maintained for a long time, so if there are any
future issues (e.g., a TF2 update breaks the HUD), it could leave users
of this project with a broken installation, which would be terrible for
those depending on the "set and forget" promise.

For this reason, idHUD is being replaced with NeoDefault's own HUD
files. Currently, these files only modify the appearance of damage
numbers, which was the main benefit of idHUD. Now, the same effect is
accomplished with a very minimal change in the default HUD.