Rework component structure to package VPKs
Instead of shipping ZIP files and unzipping those onto the user's
machine, it is more preferable to distribute these components as VPK
files. Since it's better to leave source files in the repo and build
them when needed, a build.sh script has been created to do this.

Now, components are stored in the components/ directory, and build.sh
will package these source components into the build/ directory. The
source code of the installer has also been updated to reflect this when
running the tool in debugging mode.