Add small optimization to path canon.
If you navigate between the first and second pages repeatedly, the same
path gets canonicalized over and over again. This is because
PreparePathPanel(), which needs to be called before loading the second
page, has to call SetTFPath().  To get around this, check if the
provided path is already equal to tfPath.