Fix a bug with folder selection on UI
After clicking "Select Folder", cancelling out of the menu caused the UI
to act as if an invalid path had been provided. Now, a clear distinction
is made between a failure and a cancel when calling RequestFilepath().
Furthermore, to simplify this logic, RequestFilepath() will leave the
handling of the UI elements to the caller.