Change CFG files to have CRLF when built
As a foreword-thinking measure, the CFG output files had been given
UNIX-style line endings so that there'd be easier integration for a UNIX
installer tool. Windows 10's Notepad is able to recognize these, but
Windows 8 cannot, and results in an unreadable file. Because of this,
the files will now have CRLF and any potential future issues will just
be dealt with at that time.